Added class RayIntersectionData to replace Grid's RayData list iterator.
The class RayIntersectionData contains said list iterator, but also a pointer
to the ray itself. That way we can do bounds checking, e.g., when we want to
move forward or backward along the ray via std::prev() or std::next(), we can
check that the result does not go past Ray.raydata.begin() or
Ray.raydata.end().