Ray: fixed bug in calc_SC_coeffs() where Delta_tau was rounded to zero.
This happens when the optical depths are *extremely* large, and we don't
have enough significant digits (15 or so in double precision) to store
the difference between them, and so the computer rounds the answer to
zero. Havoc ensues. The way to fix this is by lowering the maximum
optical depth.