EOS: scaled RHS of LTE root finder by n_g to make it unitless
Now we measure the change in the LTE root function relative to the total
number of particles in the voxel. Without a scaling factor, the actual
value of the root function will change with the size of the problem,
which we don't want because then it will be impossible to know if it's
actually converged.