EOS: fixed awful and subtle pointer address bug when building model atoms
It never showed up when the code was compiled with clang++. With Intel
and GCC it worked if optimization was turned off, but with any degree of
optimization it would crash due to some invalid pointers in AtomicLine
objects to AtomicLevel objects. I think it had something to do with
using push_back() with objects which have member pointers. That causes
addresses to get mangled. Now I allocate the space in the lines vector
and then insert the addresses by accessing the vector index directly.
It's not as object-y as I would like but it fixes the segfaults with GCC
and Intel.