initialize_rays(): removed a lot of crap "initialization" stuff
Our "initialization" should be setting the matter to LTE and then
solving the transfer equation. We don't need to set S directly to B or
any of that crap. There was also a temporary routine that set the
opacity to the matter density (which has since been removed). These junk
initialization steps actually affected the final result significantly,
which is scary.