ALI: size of Lambda_star set to size of grid, not only to YAML parameter
The dimension of Lambda_star was always being set to whatever
"n_depth_pts" was in the YAML file. But if we read our data from a MESA
model, that may be a different number. So just set the dimension to the
size of the grid vector so that it is correct for both toy models and
MESA models.