added function calc_emergent_spectrum()
The emergent flux is NOT the same as the flux in the outermost layer.
The former can never be negative (you can't see light moving away from
you), while the latter certainly can. So the integral over the specific
intensity to calculate the emergent flux is over only outward-pointing
rays.